(refactor) Move date validations to their own file
