+++ app/form_models/application_details_form.rb
@@ -6,6 +6,7 @@ class ApplicationDetailsForm < VacancyForm
   attr_accessor :original_publish_on
 
   include VacancyApplicationDetailValidations
+  include VacancyDateValidations
 
   def disable_editing_publish_on?
     published? && vacancy.reload.publish_on.past?
